+++ benchllm/cli/evaluator/__init__.py
@@ -1,0 +1,2 @@
+from benchllm.cli.evaluator.interactive import InteractiveEvaluator  # noqa
+from benchllm.cli.evaluator.web import WebEvaluator  # noqa
